Replace UTF-8 decoder with Björn Höhrmann's DFA implementation and update `ch_put` to handle Unicode codepoints.

+++ src/sys/chargen.h
@@ -14,7 +14,7 @@ extern void ch_update(void);
 extern void ch_pos_get(TPOS *pos);
 extern void ch_pos_set(TPOS *pos);
 extern void ch_attrib( u8 colour, u16 flags );
-extern void ch_put(u8 ch);
+extern void ch_put(u32 ch);
 extern void ch_scroll(TPOS *pos1, TPOS *pos2, s16 scroll, u8 attrib);
 extern void ch_clear(TPOS *pos1, TPOS *pos2, u8 attrib);
